Allows use of categories and tags
I probably wouldn't handle it like this if I were writing it from scratch, but this is a suggested solution for having both tags _and_ categories present on the site.

```
---
title: "Wheelie"
date: 2024-03-01T22:06:25Z
draft: false
tags: ["skoolkit", "wheelie", "zx spectrum", "z80"]
categories:
  - "Disassembly"
---
```

The problem is, categories are defined by default so you end up with an empty page; e.g. https://meik2333.com/categories/

The alternative is to remove it, which can be done

```
[Taxonomies]
  tag = "tags"
```

(If you don't define `category = "categories"` then it's removed).